A create es delete table pontra van rakva

+++ src/main/java/Server/CommandProcessor.java
@@ -4,22 +4,29 @@ import org.json.simple.parser.ParseException;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class CommandProcessor {
     private final List<Database> databases;
     private Database usedDatabase;
+    private final MongoDBManager mongoDBManager;
 
     public CommandProcessor() throws IOException, ParseException {
         databases = Json.buildDatabases();
         usedDatabase = databases.get(0);
+        mongoDBManager = new MongoDBManager();
     }
 
     public List<Database> getDatabases() {
         return databases;
     }
 
-    public void processCommand(String command) throws DatabaseExceptions.DataDefinitionException, DatabaseExceptions.UnknownCommandException, DatabaseExceptions.UnsuccesfulDeleteException, DatabaseExceptions.DatabaseNotExistsException {
+    public void processCommand(String command) throws DatabaseExceptions.DataDefinitionException, DatabaseExceptions.UnknownCommandException, DatabaseExceptions.UnsuccesfulDeleteException {
+        if(Character.compare(command.charAt(command.length() - 1),';') == 0){
+            command = command.substring(0, command.length()-1);
+        }
         String[] dividedCommand = command.split(" ");
         switch (dividedCommand[0].toUpperCase()) {
             case "CREATE":
@@ -47,7 +54,12 @@ public class CommandProcessor {
         String[] dividedCommand = command.split(" ");
         switch (dividedCommand[1].toUpperCase()){
             case "DATABASE":
-                databases.add(new Database(dividedCommand[2]));
+                if(databases.stream().noneMatch(o ->o.getName().equals(dividedCommand[2]))) {
+                    databases.add(new Database(dividedCommand[2]));
+                }
+                else{
+                    throw (new DatabaseExceptions.DataDefinitionException("database already exists: " + dividedCommand[2]));
+                }
                 break;
             case "TABLE":
                 if(usedDatabase.getTables().stream().noneMatch(o -> o.getName().equals(dividedCommand[2]))){
@@ -72,14 +84,14 @@ public class CommandProcessor {
                                 newTable.addAttribute(new Attribute(parameter));
                             }
                             else{
-                                throw (new DatabaseExceptions.DataDefinitionException("This column name already exist in this table."));
+                                throw (new DatabaseExceptions.DataDefinitionException("Column already exist in this table: " + parameter[0]));
                             }
                         }
                     }
                     usedDatabase.addTable(newTable);
                 }
                 else{
-                    throw (new DatabaseExceptions.DataDefinitionException("This tabl name already exist in this database."));
+                    throw (new DatabaseExceptions.DataDefinitionException("Table already exist in this database: " + dividedCommand[2]));
                 }
                 break;
 
@@ -88,7 +100,7 @@ public class CommandProcessor {
         }
     }
 
-    public void delete(String[] command) throws DatabaseExceptions.UnknownCommandException, DatabaseExceptions.UnsuccesfulDeleteException {
+    public void delete(String[] command) throws DatabaseExceptions.UnknownCommandException, DatabaseExceptions.DataDefinitionException {
         //DELETE DATABASE <databasename>
         //DELETE TABLE <tablename>
         switch (command[1].toUpperCase()){
@@ -96,6 +108,9 @@ public class CommandProcessor {
                 Database database = databases.stream()
                         .filter(o -> o.getName().equals(command[2]))
                         .findAny().orElse(null);
+                if(database == null){
+                    throw (new DatabaseExceptions.DataDefinitionException("Database does not exist: " + command[2]));
+                }
                 databases.remove(database);
                 break;
 
@@ -103,6 +118,9 @@ public class CommandProcessor {
                 Table table = usedDatabase.getTables().stream()
                         .filter(o -> o.getName().equals(command[2]))
                         .findAny().orElse(null);
+                if(table == null){
+                    throw (new DatabaseExceptions.DataDefinitionException("Table does not exist: " + command[2]));
+                }
                 usedDatabase.removeTable(table);
                 break;
             default:
@@ -112,15 +130,21 @@ public class CommandProcessor {
     }
 
     public void insert(String[] dividedCommand){
-
+        String parameters = dividedCommand[dividedCommand.length - 1];
+        parameters = parameters.substring(1, parameters.length() - 1);
+        String key = parameters.split(",")[0];
+        parameters = Arrays.stream(parameters.split(",")).skip(1).collect(Collectors.joining("#"));
+        System.out.println(key + " " + parameters);
+        mongoDBManager.insert(dividedCommand[2], key, parameters);
     }
 
-    public void use( String[] command) throws DatabaseExceptions.DatabaseNotExistsException {
+    public void use( String[] command) throws DatabaseExceptions.DataDefinitionException {
         usedDatabase = databases.stream()
                 .filter(o -> o.getName().equals(command[1]))
                 .findAny().orElse(null);
         if(usedDatabase == null){
-            throw (new DatabaseExceptions.DatabaseNotExistsException("This Database not exists"));
+            throw (new DatabaseExceptions.DataDefinitionException("Database does not exist: " + command[1]));
         }
+        mongoDBManager.use(usedDatabase.getName());
     }
 }
